refactor: Move `XmlLayout` conversion to a separate sub-module and dedicated traits.
