+++ src/serde/xml/mod.rs
@@ -1,7 +1,9 @@
 pub(crate) mod xml_container;
+pub(crate) mod xml_layout;
 pub(crate) mod xml_relationship;
 pub(crate) mod xml_variable;
 
 pub(crate) use xml_container::XmlContainer;
+pub(crate) use xml_layout::XmlLayout;
 pub(crate) use xml_relationship::XmlRelationship;
 pub(crate) use xml_variable::XmlVariable;
